packages/ingestkit-image/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ingestkit-image"
version = "0.1.0"
description = "VLM-based image captioning for RAG indexing"
requires-python = ">=3.10"
dependencies = [
    "ingestkit-core>=0.1.0",
    "pydantic>=2.0",
    "Pillow>=10.0",
]

[project.optional-dependencies]
ollama = ["httpx>=0.27"]
qdrant = ["qdrant-client>=1.7"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-cov",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: Unit tests (no external services)",
    "integration: Integration tests (require external services)",
]
